nexhacks-server/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
livekit-api
livekit-agents
openai